freetype: s/float/float64/.
R=adg, r2
CC=golang-dev
http://codereview.appspot.com/4040045

+++ example/freetype/main.go
@@ -21,8 +21,8 @@ import (
 var (
 	dpi      = flag.Int("dpi", 72, "screen resolution in Dots Per Inch")
 	fontfile = flag.String("fontfile", "../../luxi-fonts/luxisr.ttf", "filename of the ttf font")
-	size     = flag.Float("size", 12, "font size in points")
-	spacing  = flag.Float("spacing", 1.5, "line spacing (e.g. 2 means double spaced)")
+	size     = flag.Float64("size", 12, "font size in points")
+	spacing  = flag.Float64("spacing", 1.5, "line spacing (e.g. 2 means double spaced)")
 	wonb     = flag.Bool("whiteonblack", false, "white text on a black background")
 )
